+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 * Fixes a bug where the Cloud Functions emulator ignored the "host" configuration (#1722).
 * Fixes a bug where the Cloud Functions emulator accepted requests to too many paths (#1773).
 * Modifies `firebase ext:update` to not perform update if the extension is already up to date.
+* Print Firebase Console links for Extensions after operations.